RAB-1336: keep comment after review
