+++ src/Akeneo/Platform/Installer/back/src/Domain/CommandExecutor/DoctrineMigrationsVersionInterface.php
@@ -7,8 +7,8 @@ declare(strict_types=1);
  * @license   https://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
-namespace Akeneo\Platform\Installer\Domain\Query\CommandExecutor;
+namespace Akeneo\Platform\Installer\Domain\CommandExecutor;
 
-interface DoctrineSchemaUpdateInterface extends CommandExecutorInterface
+interface DoctrineMigrationsVersionInterface extends CommandExecutorInterface
 {
 }
