+++ src/Akeneo/Platform/Installer/back/src/Infrastructure/Query/CommandExecutor/AbstractCommandExecutor.php
@@ -9,7 +9,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Platform\Installer\Infrastructure\Query\CommandExecutor;
 
-use Akeneo\Platform\Installer\Domain\Query\CommandExecutor\CommandExecutorInterface;
+use Akeneo\Platform\Installer\Domain\CommandExecutor\CommandExecutorInterface;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Output\BufferedOutput;
